refactor(request): split enum errors between organization and application + move request helpers to shared + remove request module
