+++ backend/src/modules/application/constants/application-error.constants.ts
@@ -27,3 +27,32 @@ export const ONG_APPLICATION_ERRORS = {
     errorCode: 'ONG_APP_005',
   },
 };
+
+export const APPLICATION_REQUEST_ERRORS = {
+  CREATE: {
+    APPLICATION_STATUS: {
+      message: 'Application status must be active.',
+      errorCode: 'APP_REQ_001',
+    },
+    REQ_EXISTS: {
+      message: 'There is already a pending request with the same data.',
+      errorCode: 'APP_REQ_002',
+    },
+    APP_EXISTS: {
+      message: 'The app is already asssigned to the organization.',
+      errorCode: 'APP_REQ_003',
+    },
+  },
+  GET: {
+    NOT_FOUND: {
+      message: 'Request not found',
+      errorCode: 'APP_REQ_004',
+    },
+  },
+  UPDATE: {
+    NOT_PENDING: {
+      message: 'Could not update a Request with other status than pending.',
+      errorCode: 'APP_REQ_005',
+    },
+  },
+};
