+++ backend/src/modules/organization/constants/errors.constants.ts
@@ -70,3 +70,28 @@ export const INVESTOR_ERRORS = {
     errorCode: 'INV_001',
   },
 };
+
+export const ORGANIZATION_REQUEST_ERRORS = {
+  CREATE: {
+    USER_EXISTS: {
+      message: 'User email already exists.',
+      errorCode: 'REQ_001',
+    },
+    REQ_EXISTS: {
+      message: 'There is already a pending request with the same data.',
+      errorCode: 'REQ_002',
+    },
+  },
+  UPDATE: {
+    NOT_PENDING: {
+      message: 'Could not update a Request with other status than pending.',
+      errorCode: 'REQ_003',
+    },
+  },
+  GET: {
+    NOT_FOUND: {
+      message: 'Request not found',
+      errorCode: 'REQ_004',
+    },
+  },
+};
